requirements.txt
# sha256: G1VbiogAE0_a_jK30MtS9b2_3Qk3B8PdSExepZ8dmLc
ipaddr==2.1.11

# sha256: I9pYnJH1nLfGRNXOXfU51Eg0G9R5kX1t3pc_guJxkUc
lockfile==0.9.1

# sha256: GrguX3jFnvhJqQpscbWQknX1Eb6faWGHOHA9CC5jAs8
maxminddb==0.3.2

# sha256: r5dH7CcnQlsbCSUpdeIVAu5aPY0jXH9Jhp6xPgnM9OQ
redis==2.9.1

# sha256: xE35VNncyfV2gr-Pq_FyoRLt9T_zW7XOj4__GxmFq7E
hiredis==0.1.2

# sha256: RdsWgxKmgg_XhpfHtv0BlvOvkiyOvIhUDt0Nb6_tVU0
statsd==2.1.2

# sha256: SnUuNZOCfpNf_T__9oMU2jjXJPATCaBVYbZ3jockv3Q
pytz==2014.2

# sha256: 7shlMH6-fzKaaplFwVRTJlpEnNqvNxA0CCihk01T5Gg
python-dateutil==2.2
